Make using ProtocolBuilder the only way to create a protocol, remove unconventional set_ prefix
This makes it much easier to make backwards compatible changes because no clients can break with new protocol options.
