+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hypercore-protocol"
-version = "0.4.0-alpha.0"
+version = "0.4.0-alpha.1"
 license = "MIT OR Apache-2.0"
 description = "Replication protocol for Hypercore feeds"
 authors = ["Franz Heinzmann (Frando) <frando@unbiskant.org>"]
